Create 174. Dungeon Game
It is a game of Dungeon, where there is a dragon, who starts from top left room and the princess is on the bottom right room, the knight can only move right or down. 
Now every room has either positive or negative number by which the health of the knight increases or decreases and if the room has number 0 it has no effect.
The rule is if the health of the knight becomes 0 or less than 0, the knight dies.
So what we have to do is to find the minimum health of the knight so he can reach the princess.
Now the important trick is we aren't asking " How much health will knight have ?" 
We are asking the minimum health required for the knight to enter the room.
The right thinking is to think like a detective and start thinking in the place of princess and go backwards towards the knight and we find that what is the minimum health required to enter this room.
Let's try to learn this more easily -
[-2  -3   3]
[-5 -10   1]
[10  30  -5]  ← princess is here
The princess is in the (-5) room, where health is deducted by 5 and in the end the knight should have atleast 1 health so the knight needs 6 health in total 1 - (-5) = 6.
Now the logic is you have to take right or down easy path, you have to minus room effect and the health shouldn't be less than 1 .
The final answer will be the top-left room value, which will be our answer
We are writing the code in java
1) To get the rows and colums -
int m = dungeon.length;
int n = dungeon[0].length;
In this we have 3 by 3 dungeon that means m = 3, n = 3
2) DP table
int[][] dp = new int[m+1][n+1];
The DP table says dp[i][j], when you enter this cell, the minimum health required
m+1 and n+1 why ?
So that we can manage right and down boundry easily and we won't be needing extra if-condition.
We have to fill DP with infinity so we won't select wrong direction
for (int i = 0; i <= m; i++) {
    for (int j = 0; j <= n; j++) {
        dp[i][j] = Integer.MAX_VALUE;
    }
}
3) Base condition - 
dp[m][n - 1] = 1;
dp[m - 1][n] = 1;
Princess cell = (m-1, n-1)
We have set the right and down to 1 so that when the knight reaches the princess his health should be minimum 1.
4) Logic - 
int needed = Math.min(dp[i + 1][j], dp[i][j + 1]) - dungeon[i][j];
That means the knight will go either down or right 
Down → dp[i+1][j]
Right → dp[i][j+1]
It will choose the easy path from these two 
Math.min(. . .)
And you also have to subtract the effect of the current cell
If negative it drains more health and if positive it needs less health.
7) Health can't be less than 1 
dp[i][j] = Math.max(1, needed);
8) Final -
return dp[0][0];
This will return the top left health needed and minimum initial health needed
dp[i][j] = max(1,
    min(dp[i+1][j], dp[i][j+1]) - dungeon[i][j]
)
I use backward DP where dp[i][j] represents the minimum health required
to enter cell (i, j) so that the knight can reach the princess alive.